+++ frontend/src/app/(protected)/experiments/add/addExperimentSteps.tsx
@@ -1,4 +1,4 @@
-import { Step, Notification, ExperimentStateBase, ArmBase } from "./types";
+import { Step } from "../types";
 import AddMABArms from "./components/addMABArms";
 import AddABArms from "./components/addABArms";
 import AddNotifications from "./components/addNotifications";
@@ -13,16 +13,6 @@ const MABsteps: Step[] = [
   { name: "notifications", component: AddNotifications },
 ];
 
-interface MABArm extends ArmBase {
-  alpha_prior: number;
-  beta_prior: number;
-}
-
-interface MABExperimentState extends ExperimentStateBase {
-  arms: MABArm[];
-  notifications: Notification[];
-}
-
 // --- A/B test types and steps ---
 
 const ABsteps: Step[] = [
@@ -33,20 +23,8 @@ const ABsteps: Step[] = [
   { name: "notifications", component: AddNotifications },
 ];
 
-interface ABArm extends ArmBase {
-  mean_prior: number;
-  stdDev_prior: number;
-}
-
-interface ABExperimentState extends ExperimentStateBase {
-  arms: ABArm[];
-  notifications: Notification[];
-}
-
 // --- All steps ---
 
 const AllSteps = { mab: MABsteps, ab: ABsteps };
-type ExperimentState = MABExperimentState | ABExperimentState;
 
 export { AllSteps };
-export type { ExperimentState, Notification };
