+++ frontend/src/app/(protected)/experiments/types.ts
@@ -1,32 +1,95 @@
-interface NewArm {
+type MethodType = "mab" | "ab";
+
+interface Step {
+  name: string;
+  component: React.FC;
+}
+
+type Notification = {
+  notificationTrigger: "sample" | "time" | "event";
+  sampleSize?: number;
+  timeDays?: number;
+  eventName?: string;
+};
+
+interface ExperimentStateBase {
   name: string;
   description: string;
+  methodType: MethodType;
+}
+
+interface ArmBase {
+  name: string;
+  description: string;
+}
+
+interface StepComponentProps {
+  nextStep: () => void;
+  previousStep: () => void;
+}
+
+// ----- AB
+
+interface NewABArm extends ArmBase {
+  mean_prior: number;
+  stdDev_prior: number;
+}
+
+interface ABArm extends NewABArm {
+  arm_id: number;
+  mean_posterior: number;
+  stdDev_posterior: number;
+}
+
+interface ABExperimentState extends ExperimentStateBase {
+  arms: NewABArm[];
+  notifications: Notification[];
+}
+
+interface AB extends ABExperimentState {
+  experiment_id: number;
+  is_active: boolean;
+  arms: ABArm[];
+}
+
+// ----- MAB
+
+interface NewMABArm extends ArmBase {
   alpha_prior: number;
   beta_prior: number;
 }
 
-interface Arm extends NewArm {
+interface MABArm extends NewMABArm {
   arm_id: number;
   successes: number;
   failures: number;
 }
 
-interface NewMAB {
-  name: string;
-  description: string;
-  arms: NewArm[];
+interface MABExperimentState extends ExperimentStateBase {
+  arms: NewMABArm[];
+  notifications: Notification[];
 }
 
-interface MAB extends NewMAB {
+interface MAB extends MABExperimentState {
   experiment_id: number;
   is_active: boolean;
-  arms: Arm[];
+  arms: MABArm[];
 }
 
-interface BetaParams {
-  name: string;
-  alpha: number;
-  beta: number;
-}
+type ExperimentState = MABExperimentState | ABExperimentState;
 
-export type { Arm, MAB, NewArm, NewMAB, BetaParams };
+export type {
+  Step,
+  Notification,
+  ExperimentStateBase,
+  ArmBase,
+  StepComponentProps,
+  MethodType,
+  MABExperimentState,
+  ABExperimentState,
+  ExperimentState,
+  MAB,
+  AB,
+  MABArm,
+  ABArm,
+};
